projeto/requirements.txt
django
djangorestframework
django-cors-headers
drf-writable-nested
mysql-connector-python
mysqlclient
pytest
pytest-django
pyyaml
